requirements.txt
asn1crypto==0.22.0
Babel==2.7.0
certifi==2017.7.27.1
cffi==1.7.0
chardet==2.3.0
colorama==0.3.7
cryptography==1.7.1
diff-match-patch==20121119
distlib==0.2.4
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-bootstrap-pagination==1.7.1
django-bootstrap3==11.1.0
django-bootstrap3-datepicker==0.3.1
django-crequest==2016.3.16
django-crispy-forms==1.6.0
django-easy-select2==1.3.3
django-extensions==1.6.7
django-filter==1.1.0
django-forms-bootstrap==3.1.0
django-import-export==0.5.1
django-multiselectfield==0.1.8
Django-Select2==5.10.0
django-suit==0.2.25
django-tables2==1.2.2
enum34==1.1.6
et-xmlfile==1.0.1
html5lib==0.999999999
idna==2.5
ipaddress==1.0.17
jdcal==1.3
keyring==10.1
keyrings.alt==1.3
langdetect==1.0.7
ndg-httpsclient==0.4.2
numpy==1.13.1
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
pandas==0.20.3
pdfminer==20140328
phonenumberslite==8.10.14
Pillow==4.0.0
ply==3.9
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
pyOpenSSL==16.2.0
python-dateutil==2.6.1
python-libmagic==0.4.0
python-magic==0.4.13
pytz==2016.7
pyxdg==0.25
PyYAML==3.12
requests==2.12.2
SecretStorage==2.3.1
six==1.10.0
sqlparse==0.2.2
tablib==0.12.0
unicodecsv==0.14.1
urllib3==1.19.1
webencodings==0.5
xlrd==1.1.0
xlwt==1.3.0
